PiFinal/Python/websockets/requirements.txt
websockets
paho-mqtt
databasemanager-02.35.tar.gz
ruff